go/flow/keyhash: split the packed-key hash into its own package
The packed-key hash decides which shard or physical partition a document
routes to. Connectors that reproduce that routing in a keyed sink need the
same hash, but importing go/flow drags in the whole gazette consumer stack.

Move PackedKeyHash_HH64 and its fixed HighwayHash key into go/flow/keyhash,
whose only dependency is minio/highwayhash. The golden-vector regression
test moves with it, so it no longer needs etcd and the cgo bindings to run.
